python_services/requirements.txt
neo4j>=5.10.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
fastapi==0.99.1
uvicorn==0.22.0
fastapi
uvicorn
crewai
python-dotenv
pydantic
langchain-openai
httpx
# Removed explicit langchain/langgraph to let crewai resolve its own compatible versions
